+++ samplesheets/plugins.py
@@ -213,23 +213,23 @@ class ProjectAppPlugin(
         'igv_omit_bam': {
             'scope': SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT'],
             'type': 'STRING',
-            'label': 'BAM and CRAM files to omit from IGV sessions',
-            'default': '',
-            'placeholder': ', '.join(settings.SHEETS_IGV_OMIT_BAM),
-            'description': 'Comma separated list of BAM and CRAM file suffixes '
-            'to omit from generated IGV sessions. Overrides site-wide setting, '
-            'affects cancer and germline projects.',
+            'label': 'BAM and CRAM paths to omit from IGV sessions',
+            'default': ', '.join(settings.SHEETS_IGV_OMIT_BAM),
+            'description': 'Comma-separated list of iRODS path glob patterns '
+            'for omitting BAM and CRAM files from IGV sessions and study '
+            'shortcuts. Overrides site-wide setting, affects cancer and '
+            'germline projects. Update sheet cache after updating this value.',
             'user_modifiable': True,
         },
         'igv_omit_vcf': {
             'scope': SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT'],
             'type': 'STRING',
-            'label': 'VCF files to omit from IGV sessions',
-            'default': '',
-            'placeholder': ', '.join(settings.SHEETS_IGV_OMIT_VCF),
-            'description': 'Comma separated list of VCF file suffixes to omit '
-            'from generated IGV sessions. Overrides site-wide setting, affects '
-            'cancer and germline projects.',
+            'label': 'VCF paths to omit from IGV sessions',
+            'default': ', '.join(settings.SHEETS_IGV_OMIT_VCF),
+            'description': 'Comma-separated list of iRODS path glob patterns '
+            'for omitting VCF files from IGV sessions and study shortcuts. '
+            'Overrides site-wide setting, affects cancer and germline '
+            'projects. Update sheet cache after updating this value.',
             'user_modifiable': True,
         },
     }