+++ samplesheets/studyapps/cancer/utils.py
@@ -3,9 +3,9 @@
 import os
 
 from samplesheets.studyapps.utils import (
-    get_igv_omit_override,
+    get_igv_omit_list,
     check_igv_file_suffix,
-    check_igv_file_name,
+    check_igv_file_path,
 )
 from samplesheets.utils import get_latest_file_path
 
@@ -25,14 +25,13 @@ def get_library_file_path(assay, library_name, file_type, irods_backend, irods):
     assay_path = irods_backend.get_path(assay)
     query_path = os.path.join(assay_path, library_name)
     file_paths = []
-    override = get_igv_omit_override(assay.get_project(), file_type)
+    omit_list = get_igv_omit_list(assay.get_project(), file_type)
     try:
         obj_list = irods_backend.get_objects(irods, query_path)
         for obj in obj_list:
-            obj_name = obj['name'].lower()
             if check_igv_file_suffix(
-                obj_name, file_type
-            ) and check_igv_file_name(obj_name, file_type, override):
+                obj['name'].lower(), file_type
+            ) and check_igv_file_path(obj['path'], omit_list):
                 file_paths.append(obj['path'])
     except Exception:
         pass