+++ samplesheets/studyapps/germline/utils.py
@@ -7,8 +7,8 @@ from projectroles.plugins import get_backend_api
 
 from samplesheets.models import GenericMaterial
 from samplesheets.studyapps.utils import (
-    get_igv_omit_override,
-    check_igv_file_name,
+    get_igv_omit_list,
+    check_igv_file_path,
     check_igv_file_suffix,
 )
 from samplesheets.utils import get_index_by_header
@@ -32,7 +32,7 @@ def get_pedigree_file_path(file_type, source, study_tables):
         raise Exception('iRODS Backend not available')
 
     query_paths = []
-    override = get_igv_omit_override(source.study.get_project(), file_type)
+    omit_list = get_igv_omit_list(source.study.get_project(), file_type)
 
     def _get_val_by_index(row, idx):
         return row[idx]['value'] if idx else None
@@ -85,7 +85,7 @@ def get_pedigree_file_path(file_type, source, study_tables):
                 if (
                     obj['path'].startswith(query_path + '/')
                     and check_igv_file_suffix(obj['name'], file_type)
-                    and check_igv_file_name(obj['name'], file_type, override)
+                    and check_igv_file_path(obj['path'], omit_list)
                 ):
                     file_paths.append(obj['path'])
                     logger.debug('Added path: {}'.format(obj['path']))